move actions back to features
