--- src/actions/index.ts
@@ -1,6 +1,0 @@
-export * from "./actionAuthenticate";
-export * from "./actionCreateProblem";
-export * from "./actionLogin";
-export * from "./actionLogout";
-export * from "./actionRecalcProblemScore";
-export * from "./actionSubmit";
