+++ src/features/auth/index.ts
@@ -1,3 +1,4 @@
+export * from "./actions";
 export * from "./api";
 export * from "./components";
 export * from "./types";
