+++ src/features/problems/index.ts
@@ -1,3 +1,4 @@
+export * from "./actions";
 export * from "./components";
 export * from "./queries";
 export * from "./types";
