+++ src/features/submissions/actions/index.ts
@@ -1,0 +1,1 @@
+export * from "./actionSubmit";
