+++ tsconfig.json
@@ -23,24 +23,28 @@
         "./src/features/auth/api/*",
         "./src/features/auth/components/*",
         "./src/features/auth/queries/*",
+        "./src/features/auth/actions/*",
         "./src/features/auth/types",
       ],
       "@/features/problems/*":[
         "./src/features/problems/api/*",
         "./src/features/problems/components/*",
         "./src/features/problems/queries/*",
+        "./src/features/problems/actions/*",
         "./src/features/problems/types",
       ],
       "@/features/submissions/*":[
         "./src/features/submissions/api/*",
         "./src/features/submissions/components/*",
         "./src/features/submissions/queries/*",
+        "./src/features/submissions/actions/*",
         "./src/features/submissions/types",
       ],
       "@/features/rankings/*":[
         "./src/features/rankings/api/*",
         "./src/features/rankings/components/*",
         "./src/features/rankings/queries/*",
+        "./src/features/rankings/actions/*",
         "./src/features/rankings/types",
       ],
     },
